% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_wave.R
\name{add_wave}
\alias{add_wave}
\title{Add waves}
\usage{
add_wave(
  .x,
  date_var,
  normals_df,
  threshold = 5,
  threshold_cond = "gte",
  size = 3,
  var_name = "hw3"
)
}
\arguments{
\item{.x}{data}

\item{date_var}{date var}

\item{normals_df}{normal df}

\item{threshold}{threshold}

\item{threshold_cond}{threshold_cond}

\item{size}{size}

\item{var_name}{var_name}
}
\value{
data frame
}
\description{
Add waves
}
