% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_max_data.R
\docType{data}
\name{example_datasets}
\alias{example_datasets}
\alias{temp_max_data}
\alias{temp_min_data}
\title{Example datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 92040 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 92040 rows and 3 columns.
}
\usage{
temp_max_data

temp_min_data
}
\description{
Example datasets of climate indicators for four Brazilian municipalities with daily data, from 1961 to 2023.
}
\details{
\subsection{Dataset description}{
\itemize{
\item \code{temp_max_data} Maximum temperature (celsius degree)
\item \code{temp_min_data} Minimum temperature (celsius degree)
}
}

\subsection{Municipality codes}{
\itemize{
\item 3304557: Rio de Janeiro - RJ
\item 3303401: Nova Friburgo - RJ
\item 3136702: Juiz de Fora - MG
\item 3106200: Belo Horizonte - MG
}
}
}
\keyword{datasets}
