% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_precipitation.R
\name{summarise_precipitation}
\alias{summarise_precipitation}
\title{Compute precipitation indicators from grouped data}
\usage{
summarise_precipitation(.x, value_var, normals_df)
}
\arguments{
\item{.x}{grouped data, created with \code{dplyr::group_by()}}

\item{value_var}{name of the variable with temperature values.}

\item{normals_df}{normals data, created with \code{summarise_normal()}}
}
\value{
A tibble.
}
\description{
The function computes precipitation indicators from grouped data. Expects precipitation in millimeters (mm).
}
\details{
The dry and wet spells indicators are computed based on climatological normals, created with the \code{summarise_normal()} function and passed with the \code{normals_df} argument. Keys to join the normals data must be present (like id, year, and month)  and use the same names.

The following indicators are computed for each group.
\itemize{
\item{\code{count} Count of data points}
\item{\code{normal} Climatological normal, from \code{normals_df} argument}
\item{\code{mean} Average}
\item{\code{median} Median}
\item{\code{sd} Standard deviation}
\item{\code{se} Standard error}
\item{\code{max} Maximum value}
\item{\code{min} Minimum value}
\item{\code{p10} 10th percentile}
\item{\code{p25} 25th percentile}
\item{\code{p75} 75th percentile}
\item{\code{p90} 90th percentile}
\item{\code{rain_spells_3d} Count of rain spells occurences, with 3 or more consecutive days with rain above the climatological normal value}
\item{\code{rain_spells_5d} Count of rain spells occurences, with 5 or more consecutive days with rain above the climatological normal value}
\item{\code{p_1} Count of days with precipitation above 1mm}
\item{\code{p_5} Count of days with precipitation above 5mm}
\item{\code{p_10} Count of days with precipitation above 10mm}
\item{\code{p_50} Count of days with precipitation above 50mm}
\item{\code{p_100} Count of days with precipitation above 100mm}
\item{\code{d_3} Count of sequences of 3 days or more without precipitation}
\item{\code{d_5} Count of sequences of 5 days or more without precipitation}
\item{\code{d_10} Count of sequences of 10 days or more without precipitation}
\item{\code{d_15} Count of sequences of 15 days or more without precipitation}
\item{\code{d_20} Count of sequences of 20 days or more without precipitation}
\item{\code{d_25} Count of sequences of 25 days or more without precipitation}
}
}
\examples{
# Compute monthly normals
normals <- precipitation_data |>
  # Identify year
  dplyr::mutate(year = lubridate::year(date)) |>
  # Identify month
  dplyr::mutate(month = lubridate::month(date)) |>
  # Group by id variable, year and month
  dplyr::group_by(code_muni, year, month) |>
  summarise_normal(date_var = date, value_var = value, year_start = 1961, year_end = 1990) |>
  dplyr::ungroup()

# Compute indicators
precipitation_data |>
 # Identify year
 dplyr::mutate(year = lubridate::year(date)) |>
 # Identify month
 dplyr::mutate(month = lubridate::month(date)) |>
 # Group by id variable, year and month
 dplyr::group_by(code_muni, year, month) |>
 # Compute precipitation indicators
 summarise_precipitation(value_var = value, normals_df = normals) |>
 # Ungroup
 dplyr::ungroup()

}
