% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_rel_humidity.R
\name{summarise_rel_humidity}
\alias{summarise_rel_humidity}
\title{Compute relative humidity indicators from grouped data}
\usage{
summarise_rel_humidity(.x, value_var, normals_df)
}
\arguments{
\item{.x}{grouped data, created with \code{dplyr::group_by()}}

\item{value_var}{name of the variable with temperature values.}

\item{normals_df}{normals data, created with \code{summarise_normal()}}
}
\value{
A tibble.
}
\description{
The function computes relative humidity indicators from grouped data. Expects relative humidity in percentage.
}
\details{
The dry and wet spells indicators are computed based on climatological normals, created with the \code{summarise_normal()} function and passed with the \code{normals_df} argument. Keys to join the normals data must be present (like id, year, and month)  and use the same names.
The variables \code{ds3}, \code{ds5}, \code{ws3} and \code{ws5} must be present in the dataset. Those variables can be computed with the \code{add_wave()} function. Plase follow this function example for the correct arguments.

The following indicators are computed for each group.
\itemize{
\item{\code{count} Count of data points}
\item{\code{normal_mean} Climatological normal mean, from \code{normals_df} argument}
\item{\code{normal_p10} Climatological 10th percentile, from \code{normals_df} argument}
\item{\code{normal_p90} Climatological 90th percentile, from \code{normals_df} argument}
\item{\code{mean} Average}
\item{\code{median} Median}
\item{\code{sd} Standard deviation}
\item{\code{se} Standard error}
\item{\code{max} Maximum value}
\item{\code{min} Minimum value}
\item{\code{p10} 10th percentile}
\item{\code{p25} 25th percentile}
\item{\code{p75} 75th percentile}
\item{\code{p90} 90th percentile}
\item{\code{ds3} Count of dry spells occurences, with 3 or more consecutive days with relative humidity bellow the climatological normal value minus 10 percent}
\item{\code{ds5} Count of dry spells occurences, with 5 or more consecutive days with relative humidity bellow the climatological normal value minus 10 percent}
\item{\code{ws3} Count of wet spells occurences, with 3 or more consecutive days with relative humidity above the climatological normal value plus 10 percent}
\item{\code{ws5} Count of wet spells occurences, with 5 or more consecutive days with relative humidity above the climatological normal value plus 10 percent}
\item{\code{dry_days} Count of dry days, when the relative humidity is bellow the normal 10th percentile}
\item{\code{wet_days} Count of wet days, when the relative humidity is above the normal 90th percentile}
\item{\code{h_21_30} Count of days with relative humidity between 21\% and 30\%. Attention level}
\item{\code{h_12_20} Count of days with relative humidity between 12\% and 20\%. Alert level}
\item{\code{h_11} Count of days with relative humidity bellow 12\%. Emergence level}
}
}
