% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_solar_radiation.R
\name{summarise_solar_radiation}
\alias{summarise_solar_radiation}
\title{Compute solar radiation indicators from grouped data}
\usage{
summarise_solar_radiation(.x, value_var)
}
\arguments{
\item{.x}{grouped data, created with \code{dplyr::group_by()}}

\item{value_var}{name of the variable with temperature values.}
}
\value{
A tibble.
}
\description{
The function computes solar radiation indicators from grouped data. Expects solar radiation in MJm-2.
}
\details{
The following indicators are computed for each group.
\itemize{
\item{\code{count} Count of data points}
\item{\code{mean} Average}
\item{\code{median} Median}
\item{\code{sd} Standard deviation}
\item{\code{se} Standard error}
\item{\code{max} Maximum value}
\item{\code{min} Minimum value}
\item{\code{p10} 10th percentile}
\item{\code{p25} 25th percentile}
\item{\code{p75} 75th percentile}
\item{\code{p90} 90th percentile}
\item{\code{dark_3} Count of sequences of 3 days or more with solar radiation bellow 5}
\item{\code{dark_5} Count of sequences of 5 days or more with solar radiation bellow 5}
\item{\code{light_3} Count of sequences of 3 days or more with solar radiation above 25}
\item{\code{light_5} Count of sequences of 5 days or more with solar radiation above 25}
}
}
\examples{
# Compute indicators
solar_radiation_data |>
 # Identify year
 dplyr::mutate(year = lubridate::year(date)) |>
 # Identify month
 dplyr::mutate(month = lubridate::month(date)) |>
 # Group by id variable, year and month
 dplyr::group_by(code_muni, year, month) |>
 # Compute precipitation indicators
 summarise_solar_radiation(value_var = value) |>
 # Ungroup
 dplyr::ungroup()

}
