% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_temp.R
\name{summarise_temp}
\alias{summarise_temp}
\title{Compute temperature indicators from grouped data}
\usage{
summarise_temp(.x, value_var, normals_df, id_var)
}
\arguments{
\item{.x}{grouped data, created with \code{dplyr::group_by()}}

\item{value_var}{name of the variable with temperature values.}

\item{normals_df}{normals data, created with \code{summarise_normal()}}

\item{id_var}{id variable to join normals with temperature data}
}
\value{
A tibble.
}
\description{
The function computes several temperature indicators from grouped data.
}
\examples{
# Compute normals
normals <- temp_max_data |>
  dplyr::mutate(value = value - 273.15) |>
  dplyr::group_by(code_muni) |>
  summarise_normal(date_var = date, value_var = value, year_start = 1961, year_end = 1990) |>
  dplyr::ungroup()

temp_max_data |>
 # Kelvin to celsius
 dplyr::mutate(value = value - 273.15) |>
 # Identify year
 dplyr::mutate(year = lubridate::year(date)) |>
 # Group by municipality code and year
 dplyr::group_by(code_muni, year) |>
 # Compute temperature indicators
 summarise_temp(value_var = value, normals_df = normals, id_var = "code_muni") |>
 # Ungroup
 dplyr::ungroup()

}
