% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_temp_min.R
\name{summarise_temp_min}
\alias{summarise_temp_min}
\title{Compute minimum temperature indicators from grouped data}
\usage{
summarise_temp_min(.x, value_var, normals_df)
}
\arguments{
\item{.x}{grouped data, created with \code{dplyr::group_by()}}

\item{value_var}{name of the variable with temperature values.}

\item{normals_df}{normals data, created with \code{summarise_normal()}}
}
\value{
A tibble.
}
\description{
The function computes minimum temperature indicators from grouped data. Expects temperature in celsius degrees.
}
\details{
The cold spells indicators are computed based on climatological normals, created with the \code{summarise_normal()} function and passed with the \code{normals_df} argument. Keys to join the normals data must be present (like id, year, and month)  and use the same names.

The following indicators are computed for each group.
\itemize{
\item{\code{count} Count of data points}
\item{\code{normal_mean} Climatological normal mean, from \code{normals_df} argument}
\item{\code{normal_p10} Climatological 10th percentile, from \code{normals_df} argument}
\item{\code{normal_p90} Climatological 90th percentile, from \code{normals_df} argument}
\item{\code{mean} Average}
\item{\code{median} Median}
\item{\code{sd} Standard deviation}
\item{\code{se} Standard error}
\item{\code{max} Maximum value}
\item{\code{min} Minimum value}
\item{\code{p10} 10th percentile}
\item{\code{p25} 25th percentile}
\item{\code{p75} 75th percentile}
\item{\code{p90} 90th percentile}
\item{\code{cold_spells_3d} Count of cold spells occurences, with 3 or more consecutive days with minimum temperature bellow the climatological normal value minus 5 celsius degrees}
\item{\code{cold_spells_5d} Count of cold spells occurences, with 5 or more consecutive days with minimum temperature bellow the climatological normal value minus 5 celsius degrees}
\item{\code{cold_days} Count of cold days, when the minimum temperature is bellow the normal 10th percentile}
\item{\code{t_0} Count of days with temperatures bellow or equal to 0 celsius degrees}
\item{\code{t_5} Count of days with temperatures bellow or equal to 5 celsius degrees}
\item{\code{t_10} Count of days with temperatures bellow or equal to 10 celsius degrees}
\item{\code{t_15} Count of days with temperatures bellow or equal to 15 celsius degrees}
\item{\code{t_20} Count of days with temperatures bellow or equal to 20 celsius degrees}
}
}
\examples{
# Compute monthly normals
normals <- temp_min_data |>
  # Identify month
  dplyr::mutate(month = lubridate::month(date)) |>
  # Group by id variable and month
  dplyr::group_by(code_muni, month) |>
  summarise_normal(date_var = date, value_var = value, year_start = 1961, year_end = 1990) |>
  dplyr::ungroup()

# Compute indicators
temp_min_data |>
 # Identify year
 dplyr::mutate(year = lubridate::year(date)) |>
 # Identify month
 dplyr::mutate(month = lubridate::month(date)) |>
 # Group by id variable, year and month
 dplyr::group_by(code_muni, year, month) |>
 # Compute minimum temperature indicators
 summarise_temp_min(value_var = value, normals_df = normals) |>
 # Ungroup
 dplyr::ungroup()

}
