% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_windspeed.R
\name{summarise_windspeed}
\alias{summarise_windspeed}
\title{Compute windspeed indicators from grouped data}
\usage{
summarise_windspeed(.x, value_var, normals_df)
}
\arguments{
\item{.x}{grouped data, created with \code{dplyr::group_by()}}

\item{value_var}{name of the variable with temperature values.}

\item{normals_df}{normals data, created with \code{summarise_normal()}}
}
\value{
A tibble.
}
\description{
The function computes windspeed (u2) indicators from grouped data. Expects windspeed in kilometers per hour (km/h). 1 m/s equals to 3.6 km/h.
}
\details{
The high and low u2 indicators are computed based on climatological normals, created with the \code{summarise_normal()} function and passed with the \code{normals_df} argument. Keys to join the normals data must be present (like id, year, and month)  and use the same names.
The variables \code{l_u2_3}, \code{l_u2_5}, \code{h_u2_3} and \code{h_u2_5} must be present in the dataset. Those variables can be computed with the \code{add_wave()} function. Plase follow this function example for the correct arguments.

The following indicators are computed for each group.
\itemize{
\item{\code{count} Count of data points}
\item{\code{normal_mean} Climatological normal mean, from \code{normals_df} argument}
\item{\code{normal_p10} Climatological 10th percentile, from \code{normals_df} argument}
\item{\code{normal_p90} Climatological 90th percentile, from \code{normals_df} argument}
\item{\code{mean} Average}
\item{\code{median} Median}
\item{\code{sd} Standard deviation}
\item{\code{se} Standard error}
\item{\code{max} Maximum value}
\item{\code{min} Minimum value}
\item{\code{p10} 10th percentile}
\item{\code{p25} 25th percentile}
\item{\code{p75} 75th percentile}
\item{\code{p90} 90th percentile}
\item{\code{l_eto_3} Count of sequences of 3 days or more with windspeed bellow the climatological normal}
\item{\code{l_eto_5} Count of sequences of 5 days or more with windspeed bellow the climatological normal}
\item{\code{h_eto_3} Count of sequences of 3 days or more with windspeed above the climatological normal}
\item{\code{h_eto_5} Count of sequences of 5 days or more with windspeed above the climatological normal}
\item{\code{b1}, \code{b2}, \code{b3}, \code{b4}, \code{b5}, \code{b6}, \code{b7}, \code{b8}, \code{b9}, \code{b10}, \code{b11}, \code{b12} Beaufort scale classifications}
}
}
