% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_max_data.R
\docType{data}
\name{temp_max_data}
\alias{temp_max_data}
\title{Maximum temperature data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 106652 rows and 3 columns.
}
\usage{
temp_max_data
}
\description{
Example dataset with daily maximum temperature on Kelvin for four Brazilian municipalities (Rio de Janeiro - RJ, Nova Friburgo - RJ, Juiz de Fora - MG, Belo Horizonte - MG), from 1950 to 2022.
}
\keyword{datasets}
