% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_cut.R
\name{year_cut}
\alias{year_cut}
\title{Year cut}
\usage{
year_cut(.x, date_var = NULL, n)
}
\arguments{
\item{.x}{a tibble.}

\item{date_var}{Date variable. If \code{NULL}, the function will default to 'date'.}

\item{n}{numeric. Size of the interval, in years. Minimum of 1.}
}
\value{
a tibble with the variable \code{year_ref}.
}
\description{
Discretize dates into yearly intervals
}
\examples{
year_cut(temp_max_data, n = 30)

}
