% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_interval.R
\name{year_interval}
\alias{year_interval}
\title{Get yearly intervals}
\usage{
year_interval(x, n)
}
\arguments{
\item{x}{a date object}

\item{n}{numeric. Size of the interval in years. Minimum of 1.}
}
\value{
The starting and ending year of the interval open to the right, separated by \verb{--}.
}
\description{
Discretize dates into yearly intervals.
}
\details{
The function compute yearly intervals of size \code{n} and adds to the tibble a date interval variable (\code{date_interval}) and a interval label variable (\code{interval_label}) with the staring and ending years of the interval.
}
\examples{
year_interval(temp_max_data$date, n = 30)

}
