% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{example_datasets}
\alias{example_datasets}
\alias{temp_max_data}
\alias{temp_min_data}
\alias{rel_humidity_data}
\alias{precipitation_data}
\alias{windspeed_data}
\title{Example datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 65744 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 65744 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 65748 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 65652 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 65744 rows and 3 columns.
}
\usage{
temp_max_data

temp_min_data

rel_humidity_data

precipitation_data

windspeed_data
}
\description{
Example datasets of climate indicators for four Brazilian municipalities with daily data, from 1980 to 2025, derived from Copernicus ERA5-Land data.
}
\details{
\subsection{Dataset description}{
\itemize{
\item \code{temp_max_data} Maximum temperature (celsius degree)
\item \code{temp_min_data} Minimum temperature (celsius degree)
\item \code{rel_humidity_data} Relative humidity (percent)
\item \code{precipitation_data} Precipitation (mm)
\item \code{windspeed_data} Windspeed at 2m height (m/s)
}
}

\subsection{Municipality codes}{
\itemize{
\item 3304557: Rio de Janeiro - RJ
\item 3303401: Nova Friburgo - RJ
\item 3136702: Juiz de Fora - MG
\item 3106200: Belo Horizonte - MG
}
}
}
\keyword{datasets}
