% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_normal.R
\name{summarise_normal}
\alias{summarise_normal}
\title{Compute normals from grouped data}
\usage{
summarise_normal(.x, date_var, value_var, year_start, year_end)
}
\arguments{
\item{.x}{grouped data, created with \code{dplyr::group_by()}}

\item{date_var}{name of the variable with dates.}

\item{value_var}{name of the variable with values.}

\item{year_start}{starting year of the normal}

\item{year_end}{ending year of the normal}
}
\value{
A tibble.
}
\description{
The function computes normals (average) of a variable for each group.
}
\examples{
temp_max_data |>
 # Group by municipalities
 dplyr::group_by(code_muni) |>
 # Compute normals
 summarise_normal(date_var = date, value_var = value, year_start = 1961, year_end = 1990) |>
 # Ungroup
 dplyr::ungroup()

}
