% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_solar_radiation.R
\name{summarise_solar_radiation}
\alias{summarise_solar_radiation}
\title{Compute solar radiation indicators from grouped data}
\usage{
summarise_solar_radiation(.x, value_var, normals_df)
}
\arguments{
\item{.x}{grouped data, created with \code{dplyr::group_by()}}

\item{value_var}{name of the variable with temperature values.}

\item{normals_df}{normals data, created with \code{summarise_normal()}}
}
\value{
A tibble.
}
\description{
The function computes solar radiation indicators from grouped data. Expects solar radiation in MJm-2.
}
\details{
The dark and light days indicators are computed based on climatological normals, created with the \code{summarise_normal()} function and passed with the \code{normals_df} argument. Keys to join the normals data must be present (like id, year, and month)  and use the same names.
The variables \code{d3}, \code{d5}, \code{l3} and \code{l5} must be present in the dataset. Those variables can be computed with the \code{add_wave()} function. Plase follow this function example for the correct arguments.

The following indicators are computed for each group.
\itemize{
\item{\code{count} Count of data points}
\item{\code{normal_mean} Climatological normal mean, from \code{normals_df} argument}
\item{\code{normal_p10} Climatological 10th percentile, from \code{normals_df} argument}
\item{\code{normal_p90} Climatological 90th percentile, from \code{normals_df} argument}
\item{\code{mean} Average}
\item{\code{median} Median}
\item{\code{sd} Standard deviation}
\item{\code{se} Standard error}
\item{\code{max} Maximum value}
\item{\code{min} Minimum value}
\item{\code{p10} 10th percentile}
\item{\code{p25} 25th percentile}
\item{\code{p75} 75th percentile}
\item{\code{p90} 90th percentile}
\item{\code{d3} Count of sequences of 3 days or more with solar radiation bellow the climatological normal}
\item{\code{d5} Count of sequences of 5 days or more with solar radiation bellow the climatological normal}
\item{\code{l3} Count of sequences of 3 days or more with solar radiation above the climatological normal}
\item{\code{l5} Count of sequences of 5 days or more with solar radiation above the climatological normal}
}
}
\examples{
# Compute monthly normals
normals <- solar_radiation_data |>
  # Identify month
  dplyr::mutate(month = lubridate::month(date)) |>
  # Group by id variable and month
  dplyr::group_by(code_muni, month) |>
  summarise_normal(date_var = date, value_var = value, year_start = 1961, year_end = 1990) |>
  dplyr::ungroup()

# Compute indicators
solar_radiation_data |>
# Create wave variables
dplyr::group_by(code_muni) |>
   add_wave(
     normals_df = normals,
     threshold = 0,
     threshold_cond = "lte",
     size = 3,
     var_name = "d3"
   ) |>
   add_wave(
     normals_df = normals,
     threshold = 0,
     threshold_cond = "lte",
     size = 5,
     var_name = "d5"
   ) |>
   add_wave(
     normals_df = normals,
     threshold = 0,
     threshold_cond = "gte",
     size = 3,
     var_name = "l3"
   ) |>
   add_wave(
     normals_df = normals,
     threshold = 0,
     threshold_cond = "gte",
     size = 5,
     var_name = "l5"
   ) |>
   dplyr::ungroup() |>
 # Identify year
 dplyr::mutate(year = lubridate::year(date)) |>
 # Identify month
 dplyr::mutate(month = lubridate::month(date)) |>
 # Group by id variable, year and month
 dplyr::group_by(code_muni, year, month) |>
 # Compute solar radiation indicators
 summarise_solar_radiation(value_var = value, normals_df = normals) |>
 # Ungroup
 dplyr::ungroup()

}
