% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_windspeed.R
\name{summarise_windspeed}
\alias{summarise_windspeed}
\title{Compute windspeed indicators from grouped data}
\usage{
summarise_windspeed(.x, value_var, normals_df)
}
\arguments{
\item{.x}{grouped data, created with \code{dplyr::group_by()}}

\item{value_var}{name of the variable with temperature values.}

\item{normals_df}{normals data, created with \code{summarise_normal()}}
}
\value{
A tibble.
}
\description{
The function computes windspeed (u2) indicators from grouped data. Expects windspeed in meters per second (m/s).
}
\details{
The high and low u2 indicators are computed based on climatological normals, created with the \code{summarise_normal()} function and passed with the \code{normals_df} argument. Keys to join the normals data must be present (like id, year, and month)  and use the same names.

The following indicators are computed for each group.
\itemize{
\item{\code{count} Count of data points}
\item{\code{normal} Climatological normal, from \code{normals_df} argument}
\item{\code{mean} Average}
\item{\code{median} Median}
\item{\code{sd} Standard deviation}
\item{\code{se} Standard error}
\item{\code{max} Maximum value}
\item{\code{min} Minimum value}
\item{\code{p10} 10th percentile}
\item{\code{p25} 25th percentile}
\item{\code{p75} 75th percentile}
\item{\code{p90} 90th percentile}
\item{\code{l_eto_3} Count of sequences of 3 days or more with evapotranspirations bellow the climatological normal}
\item{\code{l_eto_5} Count of sequences of 5 days or more with evapotranspirations bellow the climatological normal}
\item{\code{h_eto_3} Count of sequences of 3 days or more with evapotranspirations above the climatological normal}
\item{\code{h_eto_5} Count of sequences of 5 days or more with evapotranspirations above the climatological normal}
}
}
\examples{
# Compute monthly normals
normals <- windspeed_data |>
  # Identify year
  dplyr::mutate(year = lubridate::year(date)) |>
  # Identify month
  dplyr::mutate(month = lubridate::month(date)) |>
  # Group by id variable, year and month
  dplyr::group_by(code_muni, year, month) |>
  summarise_normal(date_var = date, value_var = value, year_start = 1961, year_end = 1990) |>
  dplyr::ungroup()

# Compute indicators
windspeed_data |>
 # Identify year
 dplyr::mutate(year = lubridate::year(date)) |>
 # Identify month
 dplyr::mutate(month = lubridate::month(date)) |>
 # Group by id variable, year and month
 dplyr::group_by(code_muni, year, month) |>
 # Compute windspeed indicators
 summarise_windspeed(value_var = value, normals_df = normals) |>
 # Ungroup
 dplyr::ungroup()

}
